refactor: files structre for slices
