+++ src/components/NavBar/NavBar.js
@@ -14,7 +14,7 @@ import {
   PRODUCT_ROUTE,
 } from '../../route';
 import LanguageButton from '../LanguageButton/LanguageButton';
-import { Signout, user } from '../../features/user/userSlice';
+import { Signout, user } from '../../features/slices/user';
 
 const classNames = (...classes) => {
   return twMerge(classes);
