+++ src/components/SignIn/SignIn.js
@@ -13,7 +13,7 @@ import {
   signInWithGoogle,
   status,
   user,
-} from '../../features/user/userSlice';
+} from '../../features/slices/user';
 import { errorTypes } from '../../utils/errorTypes';
