+++ src/components/SignUp/SignUp.js
@@ -14,7 +14,7 @@ import {
   signUpWithCredentials,
   status,
   user,
-} from '../../features/user/userSlice';
+} from '../../features/slices/user';
 import { errorTypes } from '../../utils/errorTypes';
 
 const schema = yup.object().shape({
